Feat: Issues 상태 관리 Context 추가 (IssueContext)
